requirements.txt
# requirements.txt for WebWanderer
requests
beautifulsoup4
supabase-py
flask
gunicorn
